IHT-4971 Refactor
